Renames CredentialsAutoConfiguration to GcpPropertiesAutoConfiguration.
