+++ spring-cloud-gcp-starters/spring-cloud-gcp-starter-core/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.springframework.cloud.gcp.core.autoconfig.CredentialsAutoConfiguration,\
-  org.springframework.cloud.gcp.core.autoconfig.GcpPropertiesAutoConfiguration
+org.springframework.cloud.gcp.core.autoconfig.CredentialsAutoConfiguration
